Remove member config files and use the ones from toolchain-common
